# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "RotaTour API",
        "description": "Documentation created using Swagger.",
        "contact": {
            "name": "Saulo Gomes",
            "url": "https://github.com/saulobr88"
        },
        "version": "1.0"
    },
    "host": "rotatourapi.herokuapp.com",
    "basePath": "",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/routes": {
            "get": {
                "tags": [
                    "routes"
                ],
                "summary": "Display a listing of the routes.",
                "description": "Returns the user's routes.",
                "operationId": "api.routes.index",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Success - User found and return Routes."
                    },
                    "404": {
                        "description": "User not found."
                    }
                }
            },
            "post": {
                "tags": [
                    "routes"
                ],
                "summary": "Store a newly created route in storage.",
                "description": "Store a user's new route.",
                "operationId": "api.routes.store",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "body",
                        "description": "Name of the new Route",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewRoute"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - User found and will save the Route."
                    },
                    "404": {
                        "description": "User not found."
                    }
                }
            }
        },
        "/api/routes/show/{id}": {
            "get": {
                "tags": [
                    "routes"
                ],
                "summary": "Display the specified route.",
                "description": "Returns the route details.",
                "operationId": "api.routes.show",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Route id in database",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - Route found."
                    },
                    "404": {
                        "description": "Route not found."
                    }
                }
            }
        },
        "/api/routes/delete/{id}": {
            "delete": {
                "tags": [
                    "routes"
                ],
                "summary": "Remove the specified route from storage.",
                "description": "Delete the specified route.",
                "operationId": "api.routes.delete",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Route id in database",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - Route found."
                    },
                    "404": {
                        "description": "Route not found."
                    }
                }
            }
        },
        "/api/routes/addToRoute": {
            "get": {
                "tags": [
                    "routes"
                ],
                "summary": "Add a item into a route.",
                "description": "Adds an Item to specified route.",
                "operationId": "api.routes.addToRoute",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "routeId",
                        "in": "body",
                        "description": "Route id in database",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewRoute"
                        },
                        "type": "string"
                    },
                    {
                        "name": "google_place_id",
                        "in": "body",
                        "description": "Google place - id",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewPlace"
                        },
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "body",
                        "description": "Google place - name",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewPlace"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - Add to Route."
                    },
                    "404": {
                        "description": "Route not found."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "NewPlace": {
            "required": [
                "google_place_id",
                "name"
            ],
            "properties": {
                "google_place_id": {
                    "description": "UUID Google Place Id",
                    "type": "string",
                    "example": "ChIJVyuijGQZqwcREEzZ32LILvA"
                },
                "name": {
                    "description": "Place name",
                    "type": "string",
                    "example": "Hospital das Cl\u00ednicas - HC - UFPE"
                }
            }
        },
...
        "NewRoute": {
            "required": [
                "name",
                "user_id"
            ],
            "properties": {
                "name": {
                    "description": "Route's full name",
                    "type": "string",
                    "example": "Rota de teste"
                },
                "user_id": {
                    "description": "User's id in database",
                    "type": "integer"
                }
            }
        },
...
    }
}